+++ cmd/rove-server/internal/accounts_test.go
@@ -1,4 +1,4 @@
-package accounts
+package internal
 
 import (
 	"testing"
