+++ cmd/rove-server/internal/simpleAccountant.go
@@ -1,4 +1,4 @@
-package accounts
+package internal
 
 import (
 	"fmt"
